LayoutTests/accessibility/mac/nested-inline-elements-children.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script>
    function dumpAccessibilityChildren(element, level) {
        if (element.stringValue.indexOf('End of test') >= 0)
            return false;

        var indent = "";
        for (var k = 0; k < level; k++) { indent += "  "; }
        debug(indent + element.role + " " + element.stringValue);
        var childrenCount = element.childrenCount;
        for (var k = 0; k < childrenCount; k++) {
            if (!dumpAccessibilityChildren(element.childAtIndex(k), level+1))
                return false;
        }
        return true;
    }
</script>
</head>

<body id="body">

<div>
<span>
    <span>
        test1
        <div></div>
        <span> test2</span>
    </span><div>
    test3
    </div>
</span>
</div>

<p id="description"></p>
<div id="console"></div>
<script>
    if (window.accessibilityController) {
        description("This tests that when you have nested inline continuations, we add the children correctly.");

        var content = accessibilityController.rootElement.childAtIndex(0).childAtIndex(0);
        dumpAccessibilityChildren(content, 0);
        
        shouldBe("content.childrenCount", "3");
    }
    successfullyParsed = true;
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
