Safari-611.1.10/LayoutTests/accessibility/mac/nested-inline-elements-children-expected.txt

test1
test2
test3
This tests that when you have nested inline continuations, we add the children correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


AXRole: AXGroup AXValue:
  AXRole: AXStaticText AXValue: test1
  AXRole: AXStaticText AXValue: test2
  AXRole: AXGroup AXValue:
    AXRole: AXStaticText AXValue: test3
PASS content.childrenCount is 3
PASS successfullyParsed is true

TEST COMPLETE
